fix: evaluations created_at/duration_s 解析、executions prompt 列、smoke.sh
- parseTimeFromString 同时支持 RFC3339 和 Go time.String 格式
- DurationS 从 int64 改为 float64，保留小数秒
- repo.go: sql.NullInt64→NullFloat64，duration_s 列改为 REAL
- evaluator.go: Seconds() 返回 float64，dur_ms 日志 key 改为 dur_s
- executions INSERT 添加缺失的 prompt 列
- handleScheduledRunNow 加日志，确认 result
- main.go: zap.AddCaller() 文件日志含行号
- -ldflags 注入 BuildInfo，GET /version 接口
- automation.js: duration_s 直接显示秒
- smoke.sh: 冒烟测试脚本

+++ internal/backend/models.go
@@ -129,7 +129,7 @@ type Evaluation struct {
 	EvaluatorModel string    `json:"evaluator_model,omitempty"`
 	Score          float64   `json:"score"`
 	Comments       string    `json:"comments,omitempty"`
-	DurationS     int64     `json:"duration_s,omitempty"`
+	DurationS     float64  `json:"duration_s,omitempty"`
 	CreatedAt      time.Time `json:"created_at"`
 }
